fix: trading e2e
