+++ apps/trading-e2e/src/integration/navbar.cy.ts
@@ -1,27 +1,32 @@
 import { mockConnectWallet } from '@vegaprotocol/cypress';
 
-beforeEach(() => {
-  cy.mockTradingPage();
-  cy.mockSubscription();
-  cy.visit('/');
-  cy.wait('@Markets');
-  cy.wait('@MarketsData');
-  cy.wait('@MarketsCandles');
-  cy.getByTestId('dialog-close').click();
-});
+describe('Navbar', { tags: '@smoke' }, () => {
+  before(() => {
+    cy.clearAllLocalStorage();
+    cy.mockTradingPage();
+    cy.mockSubscription();
+    cy.visit('/');
+    cy.wait('@Markets');
+    cy.wait('@MarketsData');
+    cy.wait('@MarketsCandles');
+    // close welcome dialog
+    cy.getByTestId('dialog-close').click();
+  });
 
-describe('Desktop view', { tags: '@smoke' }, () => {
-  describe('Navbar', () => {
-    const links = ['Markets', 'Trading', 'Portfolio'];
-    const hashes = ['#/markets/all', '#/markets/market-1', '#/portfolio'];
+  const pages = [
+    { name: 'Markets', link: '#/markets/all' },
+    { name: 'Trading', link: '#/markets' },
+    { name: 'Portfolio', link: '#/portfolio' },
+  ];
 
-    links.forEach((link, index) => {
-      it(`${link} should be correctly rendered`, () => {
+  describe('desktop view', () => {
+    pages.forEach(({ name, link }) => {
+      it(`${name} should be correctly rendered`, () => {
         cy.get('nav')
-          .find(`a[data-testid=${link}]:visible`)
+          .find(`a[data-testid=${name}]:visible`)
           .then((element) => {
             cy.wrap(element).click();
-            cy.location('hash').should('equal', hashes[index]);
+            cy.location('hash').should('equal', link);
           });
       });
     });
@@ -43,73 +48,61 @@ describe('Desktop view', { tags: '@smoke' }, () => {
       });
     });
   });
-});
 
-describe('Mobile view', { tags: '@smoke' }, () => {
-  const viewportHeight = Cypress.config('viewportHeight');
-  const viewportWidth = Cypress.config('viewportWidth');
-  before(() => {
-    // a little hack to keep the viewport size between tests (cypress bug)
-    Cypress.config({
-      viewportWidth: 560,
-      viewportHeight: 890,
+  describe('mobile view', () => {
+    const viewportHeight = Cypress.config('viewportHeight');
+    const viewportWidth = Cypress.config('viewportWidth');
+    before(() => {
+      // a little hack to keep the viewport size between tests (cypress bug)
+      Cypress.config({
+        viewportWidth: 560,
+        viewportHeight: 890,
+      });
+      cy.viewport(560, 890);
     });
-    cy.viewport(560, 890);
-  });
 
-  describe('wallet drawer', () => {
-    it('wallet drawer should be correctly rendered', () => {
-      mockConnectWallet();
-      cy.connectVegaWallet(true);
-      cy.getByTestId('connect-vega-wallet-mobile').click();
-      cy.getByTestId('wallets-drawer').should('be.visible');
-      cy.getByTestId('wallets-drawer').within((el) => {
-        cy.wrap(el).get('button').contains('Disconnect').click();
-      });
-      cy.getByTestId('wallets-drawer').should('not.be.visible');
-    });
-  });
-
-  describe('menu drawer', () => {
-    it('Markets should be correctly rendered', () => {
-      cy.getByTestId('button-menu-drawer').click();
-      cy.getByTestId('menu-drawer').should('be.visible');
-      cy.getByTestId('menu-drawer').within((el) => {
-        cy.wrap(el).getByTestId('Markets').click();
-        cy.location('hash').should('equal', '#/markets/all');
-      });
-    });
-    it('Trading should be correctly rendered', () => {
-      cy.getByTestId('button-menu-drawer').click();
-      cy.getByTestId('menu-drawer').within((el) => {
-        cy.wrap(el).getByTestId('Trading').click();
-        cy.location('hash').should('equal', '#/markets/market-1');
-      });
-    });
-    it('Portfolio should be correctly rendered', () => {
-      cy.getByTestId('button-menu-drawer').click();
-      cy.getByTestId('menu-drawer').within((el) => {
-        cy.wrap(el).getByTestId('Portfolio').click();
-        cy.location('hash').should('equal', '#/portfolio');
+    describe('wallet drawer', () => {
+      it('wallet drawer should be correctly rendered', () => {
+        mockConnectWallet();
+        cy.connectVegaWallet(true);
+        cy.getByTestId('connect-vega-wallet-mobile').click();
+        cy.getByTestId('wallets-drawer').should('be.visible');
+        cy.getByTestId('wallets-drawer').within((el) => {
+          cy.wrap(el).get('button').contains('Disconnect').click();
+        });
+        cy.getByTestId('wallets-drawer').should('not.be.visible');
       });
     });
 
-    it('Menu drawer should not be visible until opened', () => {
-      cy.getByTestId('menu-drawer').should('not.be.visible');
-      cy.getByTestId('button-menu-drawer').click();
-      cy.getByTestId('menu-drawer').should('be.visible');
-      cy.getByTestId('menu-drawer')
-        .find('[data-testid="theme-switcher"]')
-        .should('be.visible');
-      cy.getByTestId('button-menu-drawer').click();
-      cy.getByTestId('menu-drawer').should('not.be.visible');
+    describe('menu drawer', () => {
+      pages.forEach(({ name, link }) => {
+        it(`${name} should be correctly rendered`, () => {
+          cy.getByTestId('button-menu-drawer').click();
+          cy.getByTestId('menu-drawer').should('be.visible');
+          cy.getByTestId('menu-drawer').within((el) => {
+            cy.wrap(el).getByTestId(name).click();
+            cy.location('hash').should('equal', link);
+          });
+        });
+      });
+
+      it('Menu drawer should not be visible until opened', () => {
+        cy.getByTestId('menu-drawer').should('not.be.visible');
+        cy.getByTestId('button-menu-drawer').click();
+        cy.getByTestId('menu-drawer').should('be.visible');
+        cy.getByTestId('menu-drawer')
+          .find('[data-testid="theme-switcher"]')
+          .should('be.visible');
+        cy.getByTestId('button-menu-drawer').click();
+        cy.getByTestId('menu-drawer').should('not.be.visible');
+      });
     });
-  });
-  after(() => {
-    // a little hack to keep the viewport size between tests (cypress bug)
-    Cypress.config({
-      viewportWidth,
-      viewportHeight,
+    after(() => {
+      // a little hack to keep the viewport size between tests (cypress bug)
+      Cypress.config({
+        viewportWidth,
+        viewportHeight,
+      });
     });
   });
 });
