+++ apps/trading-e2e/src/integration/trading-accounts.cy.ts
@@ -2,6 +2,7 @@ import { checkSorting } from '@vegaprotocol/cypress';
 
 describe('accounts', { tags: '@smoke' }, () => {
   before(() => {
+    cy.clearAllLocalStorage();
     cy.mockTradingPage();
     cy.mockWeb3Provider();
     cy.mockSubscription();
@@ -42,13 +43,21 @@ describe('accounts', { tags: '@smoke' }, () => {
       .should('have.text', '100,001.01');
   });
 
-  it('asset detail should be properly rendered', () => {
+  it('should open asset details dialog when clicked on symbol', () => {
     cy.getByTestId('asset').contains('tEURO').click();
-    cy.get('[data-testid$="_label"]').should('have.length', 16);
+    cy.get('[data-testid="dialog-content"]:visible').should('exist');
     cy.get('[data-testid="dialog-close"]:visible').click();
   });
 
   describe('sorting by ag-grid columns should work well', () => {
+    before(() => {
+      const dialogs = Cypress.$('[data-testid="dialog-close"]:visible');
+      if (dialogs.length > 0) {
+        dialogs.each((btn) => {
+          cy.wrap(btn).click();
+        });
+      }
+    });
     it('sorting by asset', () => {
       cy.getByTestId('Collateral').click();
       const marketsSortedDefault = ['tBTC', 'tEURO', 'tDAI', 'tBTC'];
